implemented backup system
a backup is created when the first child is added, if any changes occure (hasChanged method is called) the backup panel offers to revert back to the backup or propagate to the children.
Later will be added the possibility to delete the children or simply create a copy wich has the modification while keeping the current bamItem untouched.