Create FixedTextAnnotation.java
Class that can be used to add text (using plot.addAnnotation) in plot region with the location unrelated to the data domain and range.

Provided position is interpreted in the java 2D space, relatively to the top left corner of the plot region.

It extends a JLabel so all formatting associated with JLabel is possible. Among other things, html is supported.