remove resetting main plot editor defaults
this introduces a small bug: when clicking "reset plot" in the plot editor, the plot is reset ignoring the current color palette. the user needs to reselect the color palette manually. No easy fix. Previous implementation did not have this issue but clicking "reset plot" would not restore  changes others than the colors such as size and shape of a point...